+++ CHANGELOG.md
@@ -13,8 +13,11 @@ First beta release with multi-label functionality and stopping criteria.
 
 - Documentation has been overhauled considerably.
 - `PoolBasedActiveLearner`: Renamed `incremental_training` kwarg to `reuse_model`.
+- `SklearnClassifier`: Changed `__init__(clf)` to `__init__(model, num_classes, multi_Label=False)`
+- `SklearnClassifierFactory`: `__init__(clf_template, kwargs={})` to `__init__(base_estimator, num_classes, kwargs={})`.
+- Refactored `KimCNNClassifier` and `TransformerBasedClassification`.
 
 ### Removed
 
 - Removed `device` kwarg from `PytorchDataset.__init__()`, 
-`PytorchTextClassificationDataset.__init__()` and `TransformersDataset.__init__()`
+`PytorchTextClassificationDataset.__init__()` and `TransformersDataset.__init__()`.